requirement.txt
python-telegram-bot==20.7
Pillow
requests
apscheduler
